# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * MediCare HMS API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface LoginBody {
  username: string;
  password: string;
  role?: string;
}

export interface RegisterBody {
  username: string;
  password: string;
  email: string;
  role: string;
  fullName: string;
}

export interface User {
  id: number;
  username: string;
  email: string;
  role: string;
  fullName: string;
  createdAt: string;
}

export interface LoginResponse {
  token: string;
  user: User;
}

export interface DashboardStats {
  totalPatients: number;
  totalDoctors: number;
  todayAppointments: number;
  totalRevenue: number;
  pendingBills: number;
  totalPrescriptions: number;
}

export interface ChartDataPoint {
  date: string;
  count: number;
  scheduled: number;
  completed: number;
  cancelled: number;
}

export interface StatusCount {
  status: string;
  count: number;
}

export interface ChartData {
  appointmentsByDay: ChartDataPoint[];
  statusBreakdown: StatusCount[];
}

export interface Department {
  id: number;
  name: string;
}

export interface Patient {
  id: number;
  /** @nullable */
  userId?: number | null;
  fullName: string;
  /** @nullable */
  dob?: string | null;
  /** @nullable */
  gender?: string | null;
  /** @nullable */
  bloodGroup?: string | null;
  /** @nullable */
  phone?: string | null;
  /** @nullable */
  address?: string | null;
  createdAt: string;
}

export interface PatientListResponse {
  patients: Patient[];
  total: number;
  page: number;
  totalPages: number;
}

export interface CreatePatientBody {
  fullName: string;
  dob?: string;
  gender?: string;
  bloodGroup?: string;
  phone?: string;
  address?: string;
  email?: string;
}

export interface UpdatePatientBody {
  fullName?: string;
  dob?: string;
  gender?: string;
  bloodGroup?: string;
  phone?: string;
  address?: string;
}

export interface DoctorDetail {
  id: number;
  /** @nullable */
  userId?: number | null;
  /** @nullable */
  departmentId?: number | null;
  fullName: string;
  /** @nullable */
  specialization?: string | null;
  /** @nullable */
  experienceYrs?: number | null;
  isAvailable: boolean;
  /** @nullable */
  departmentName?: string | null;
  /** @nullable */
  email?: string | null;
}

export interface CreateDoctorBody {
  fullName: string;
  specialization: string;
  departmentId?: number;
  experienceYrs?: number;
  email?: string;
  password?: string;
}

export interface UpdateDoctorAvailabilityBody {
  isAvailable: boolean;
}

export interface AppointmentDetail {
  id: number;
  patientId: number;
  doctorId: number;
  appointmentDate: string;
  timeSlot: string;
  status: string;
  /** @nullable */
  notes?: string | null;
  patientName: string;
  doctorName: string;
  /** @nullable */
  departmentName?: string | null;
  createdAt: string;
}

export interface AppointmentListResponse {
  appointments: AppointmentDetail[];
  total: number;
  page: number;
  totalPages: number;
}

export interface CreateAppointmentBody {
  patientId: number;
  doctorId: number;
  appointmentDate: string;
  timeSlot: string;
  notes?: string;
}

export interface TodayCountResponse {
  count: number;
}

export interface PrescriptionDetail {
  id: number;
  /** @nullable */
  appointmentId?: number | null;
  doctorId: number;
  patientId: number;
  diagnosis: string;
  medicines: string;
  /** @nullable */
  dosageInstructions?: string | null;
  issuedDate: string;
  patientName: string;
  doctorName: string;
}

export interface CreatePrescriptionBody {
  appointmentId?: number;
  doctorId: number;
  patientId: number;
  diagnosis: string;
  medicines: string;
  dosageInstructions?: string;
}

export interface BillDetail {
  id: number;
  patientId: number;
  /** @nullable */
  appointmentId?: number | null;
  consultationFee: number;
  medicineCost: number;
  totalAmount: number;
  paymentStatus: string;
  billDate: string;
  patientName: string;
  /** @nullable */
  doctorName?: string | null;
}

export interface CreateBillBody {
  patientId: number;
  appointmentId?: number;
  consultationFee: number;
  medicineCost: number;
}

export interface BillStats {
  totalBilled: number;
  totalCollected: number;
  pendingAmount: number;
  totalBills: number;
  paidBills: number;
  pendingBills: number;
}

export type GetPatientsParams = {
  search?: string;
  page?: number;
  limit?: number;
};

export type GetDoctorsParams = {
  department?: string;
};

export type GetDoctorSlotsParams = {
  date: string;
};

export type GetDoctorAppointmentsParams = {
  date?: string;
};

export type GetAppointmentsParams = {
  date?: string;
  status?: string;
  page?: number;
  limit?: number;
};

export type GetPrescriptionsParams = {
  patientId?: number;
  doctorId?: number;
};

export type GetBillsParams = {
  patientId?: number;
  status?: string;
};