# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: MediCare HMS API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginBody"
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LoginResponse"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/register:
    post:
      operationId: register
      tags: [auth]
      summary: Register new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterBody"
      responses:
        "201":
          description: User created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LoginResponse"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current user
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Current user info
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"

  # Dashboard
...
  /appointments:
    get:
      operationId: getAppointments
      tags: [appointments]
      summary: List all appointments
      security:
        - bearerAuth: []
      parameters:
        - name: date
          in: query
          schema:
            type: string
        - name: status
          in: query
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: integer
        - name: limit
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: List of appointments
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AppointmentListResponse"

    post:
      operationId: createAppointment
      tags: [appointments]
      summary: Book a new appointment
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateAppointmentBody"
      responses:
        "201":
          description: Appointment booked
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AppointmentDetail"
  /appointments/today:
    get:
      operationId: getTodayAppointmentsCount
      tags: [appointments]
      summary: Get today's appointment count
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Today's count
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TodayCountResponse"
  /appointments/{id}/cancel:
    put:
      operationId: cancelAppointment
      tags: [appointments]
      summary: Cancel an appointment
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Appointment cancelled
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AppointmentDetail"
  /appointments/{id}/complete:
    put:
      operationId: completeAppointment
      tags: [appointments]
      summary: Complete an appointment
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Appointment completed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AppointmentDetail"

  # Prescriptions
...
  /bills:
    get:
      operationId: getBills
      tags: [bills]
      summary: List all bills
      security:
        - bearerAuth: []
      parameters:
        - name: patientId
          in: query
          schema:
            type: integer
        - name: status
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of bills
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/BillDetail"

    post:
      operationId: createBill
      tags: [bills]
      summary: Generate a bill
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateBillBody"
      responses:
        "201":
          description: Bill generated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BillDetail"
  /bills/stats:
    get:
      operationId: getBillStats
      tags: [bills]
      summary: Get billing statistics
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Bill statistics
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BillStats"
  /bills/{id}/pay:
    put:
      operationId: payBill
      tags: [bills]
      summary: Mark bill as paid
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Bill paid
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BillDetail"
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
    LoginBody:
      type: object
      properties:
        username:
          type: string
        password:
          type: string
        role:
          type: string
      required:
        - username
        - password
    RegisterBody:
      type: object
      properties:
        username:
          type: string
        password:
          type: string
        email:
          type: string
        role:
          type: string
        fullName:
          type: string
      required:
        - username
        - password
        - email
        - role
        - fullName
    User:
      type: object
      properties:
        id:
          type: integer
        username:
          type: string
        email:
          type: string
        role:
          type: string
        fullName:
          type: string
        createdAt:
          type: string
      required:
        - id
        - username
        - email
        - role
        - fullName
        - createdAt
    LoginResponse:
      type: object
      properties:
        token:
          type: string
        user:
          $ref: "#/components/schemas/User"
      required:
        - token
        - user
...
    AppointmentDetail:
      type: object
      properties:
        id:
          type: integer
        patientId:
          type: integer
        doctorId:
          type: integer
        appointmentDate:
          type: string
        timeSlot:
          type: string
        status:
          type: string
        notes:
          type: ["string", "null"]
        patientName:
          type: string
        doctorName:
          type: string
        departmentName:
          type: ["string", "null"]
        createdAt:
          type: string
      required:
        - id
        - patientId
        - doctorId
        - appointmentDate
        - timeSlot
        - status
        - patientName
        - doctorName
        - createdAt
    AppointmentListResponse:
      type: object
      properties:
        appointments:
          type: array
          items:
            $ref: "#/components/schemas/AppointmentDetail"
        total:
          type: integer
        page:
          type: integer
        totalPages:
          type: integer
      required:
        - appointments
        - total
        - page
        - totalPages
    CreateAppointmentBody:
      type: object
      properties:
        patientId:
          type: integer
        doctorId:
          type: integer
        appointmentDate:
          type: string
        timeSlot:
          type: string
        notes:
          type: string
      required:
        - patientId
        - doctorId
        - appointmentDate
        - timeSlot
    TodayCountResponse:
      type: object
      properties:
        count:
          type: integer
      required:
        - count
...
    BillDetail:
      type: object
      properties:
        id:
          type: integer
        patientId:
          type: integer
        appointmentId:
          type: ["integer", "null"]
        consultationFee:
          type: number
        medicineCost:
          type: number
        totalAmount:
          type: number
        paymentStatus:
          type: string
        billDate:
          type: string
        patientName:
          type: string
        doctorName:
          type: ["string", "null"]
      required:
        - id
        - patientId
        - consultationFee
        - medicineCost
        - totalAmount
        - paymentStatus
        - billDate
        - patientName
    CreateBillBody:
      type: object
      properties:
        patientId:
          type: integer
        appointmentId:
          type: integer
        consultationFee:
          type: number
        medicineCost:
          type: number
      required:
        - patientId
        - consultationFee
        - medicineCost
    BillStats:
      type: object
      properties:
        totalBilled:
          type: number
        totalCollected:
          type: number
        pendingAmount:
          type: number
        totalBills:
          type: integer
        paidBills:
          type: integer
        pendingBills:
          type: integer
      required:
        - totalBilled
        - totalCollected
        - pendingAmount
        - totalBills
        - paidBills
        - pendingBills