# lib/api-zod/src/generated/types/appointmentListResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * MediCare HMS API specification
 * OpenAPI spec version: 0.1.0
 */
import type { AppointmentDetail } from "./appointmentDetail";

export interface AppointmentListResponse {
  appointments: AppointmentDetail[];
  total: number;
  page: number;
  totalPages: number;
}